docs: address radio button review feedback (refs SFKUI-8123)
