+++ packages/vue/src/components/FRadioField/examples/FRadioFieldDetailsAlwaysExample.vue
@@ -6,19 +6,15 @@ const illness = ref<string>();
 </script>
 
 <template>
-    <div class="row">
-        <div class="col col--md-9">
-            <f-fieldset name="illness" show-details="always">
-                <template #label> Vad hade barnet? </template>
-                <f-radio-field v-model="illness" value="temporary">
-                    Tillfällig sjukdom
-                    <template #details> som förkylning eller brutet ben </template>
-                </f-radio-field>
-                <f-radio-field v-model="illness" value="long-term">
-                    Långvarig sjukdom
-                    <template #details> som diabetes eller astma </template>
-                </f-radio-field>
-            </f-fieldset>
-        </div>
-    </div>
+    <f-fieldset name="illness" show-details="always">
+        <template #label> Vad hade barnet? </template>
+        <f-radio-field v-model="illness" value="temporary">
+            Tillfällig sjukdom
+            <template #details> som förkylning eller brutet ben </template>
+        </f-radio-field>
+        <f-radio-field v-model="illness" value="long-term">
+            Långvarig sjukdom
+            <template #details> som diabetes eller astma </template>
+        </f-radio-field>
+    </f-fieldset>
 </template>
